+++ apps/api/src/app/app.controller.ts
@@ -1,12 +1,14 @@
 import { Controller, Get } from '@nestjs/common';
+import { SkipThrottle } from '@nestjs/throttler';
 import { AppService } from './app.service';
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
 
+  @SkipThrottle()
   @Get('health')
-  getHealth() {
+  async getHealth() {
     return this.appService.getHealth();
   }